Fix test framework and pipeline
