+++ sdk/web-pubsub/web-pubsub-chat/test/public/utils/recordedClient.ts
@@ -1,8 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { Recorder, type VitestTestContext } from "@azure-tools/test-recorder";
-import recorderOptions from "../../testEnv.js";
+import {
+  Recorder,
+  type RecorderStartOptions,
+  type VitestTestContext,
+} from "@azure-tools/test-recorder";
+
+const endpoint = "https://endpoint";
+const apiKey = "api_key";
+
+const recorderOptions: RecorderStartOptions = {
+  envSetupForPlayback: {
+    WPS_CHAT_ENDPOINT: endpoint,
+    WPS_CHAT_CONNECTION_STRING: `Endpoint=${endpoint};AccessKey=${apiKey};Version=1.0;`,
+  },
+  removeCentralSanitizers: [
+    "AZSDK3430", // $..id
+    "AZSDK3433", // $..userId
+    "AZSDK3442", // $..createdBy
+    "AZSDK3490", // $..etag values are required for conditional request playback
+    "AZSDK3493", // $..name
+  ],
+};
 
 /**
  * Creates and starts the recorder for a test context.
