--- sdk/web-pubsub/web-pubsub-chat/test/utils/injectables.ts
@@ -1,21 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { inject } from "vitest";
-import { EnvVarKeys } from "./constants.js";
-
-export function getEndpoint(): string {
-  return inject(EnvVarKeys.ENDPOINT);
-}
-
-export function getConnectionString(): string {
-  return inject(EnvVarKeys.CONNECTION_STRING);
-}
-
-export function getKey(): string {
-  return inject(EnvVarKeys.API_KEY);
-}
-
-export function isLiveMode(): boolean {
-  return inject(EnvVarKeys.TEST_MODE) === "live";
-}
